JAVA-2815: Add support for transactions to the synchronous API
A few notes on the design and implementation:

* It introduces a small but significant breaking change to the existing
API for any applications that already depend on session support
(introduced in the 3.6 release to support causal consistency): the type
of ClientSession changes from com.mongodb.session.ClientSession to
com.mongodb.client.ClientSession.  This is both source and binary
incompatible.
* It adds an overload to MongoClient.startSession that takes no
ClientSessionOptions, as we expect it to be a common pattern to not
need any options.
* Internally, it moves ReadConcern from a property of ReadOperation
implementation classes to a property of the SessionContext, so that
ReadConcern can be applied properly when in a transaction (transactions
require that read concern is specified only on the first command of
a transaction, regardless of whether it's a read or a write