Deprecate API elements in driver-core module
Deprecate API element in driver-core module that are not exposed via
the high-level driver APIS (driver-sync, driver-sync, driver-legacy).
The driver-core module is a useful abstraction on which to build these
high-level drivers, but as a public API it has not seen the sort of
adoption that we initially expected.  Formally deprecating it and using
it just as an internal abstraction will allow us to more rapidly evolve
its API without concern for backwards compatibility that doesn't
actually help external users.

JAVA-3007