Fix race condition in DefaultConnectionPool constructor
The race condition is due to the constructor starting a thread which
access final fields of the containing class.  Since the thread may
execute before the constructor completes, there is no guarantee that
the final fields have been "published" to main memory, and therefore
the fields set in the constructor may not be visible to the thread
immediately.

The fix is to defer starting the thread to a start method that the
driver should call immediately after construction.

JAVA-3206