Remove JUL logging support
The Java driver has supported logging with JUL (java.util.logging) since
its inception.  Since then, SLF4J has become the widely accepted logging
standard for Java and JUL has fallen out of favor. But for backwards
compatibility the driver has continued to fall back to JUL when SLF4J
classes are not found. This commit removes that fallback, making SLF4J
the only supported logging framework for the driver. If the
"org.sl4j.Logger" class is not found, the driver will print a single
warning to JUL and not log anything else for the duration of the
process.

JAVA-2953