Add structured logging support for command logging
Add command logging unified test

Note also: I removed the "reply" from the command succeeded log message assertion.
The reason is that in the Java driver we decided that it was expensive and mostly
low value to log the reply to the command, since in the great majority of scenarios
where people want to see these log messages, they care about what the driver requested
but not what the server replied with (since for the most part they can figure that out
afterwards based on the request).

Note also: I added an appName uriOption as that made it easier to correlate log messages with clients